Shrink diff vs master's Agda.Interaction.Base
By promoting TCM-aware defs rather than demoting the rest
