+++ src/full/Agda/Interaction/JSONTop.hs
@@ -15,8 +15,8 @@ import qualified Data.Set as Set
 import Agda.Interaction.AgdaTop
 import Agda.Interaction.Base
          ( CommandState(..), CurrentFile(..), ComputeMode(..), Rewrite(..)
-         , OutputForm, OutputForm_boot(..)
-         , OutputConstraint, OutputConstraint_boot(..) )
+         , OutputConstraint_boot(..), OutputForm_boot(..))
+import Agda.Interaction.Output (OutputConstraint, OutputForm)
 import qualified Agda.Interaction.BasicOps as B
 import Agda.Interaction.EmacsTop
 import Agda.Interaction.JSON
